move get_partitions to kafka_util
